# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Verde API
  version: 0.1.0
  description: |
    API del sistema de gestión de áreas verdes (PUCP / DP2).
    Monolito modular Go + Gin + GORM. El mapa es un cliente de estas rutas;
    no hay un segundo backend geo ni GeoServer.

    CRS de las geometrías: **EPSG:4326** (longitud, latitud), GeoJSON RFC 7946.
    La columna `geom` admite NULL (catastro progresivo, sin GPS).

    Fases A–F: salud, catastro, labores, relieve y overlays de inventario.
    La sesión de la PWA es una cookie HttpOnly de cuentas locales.
    Si la cookie viene, el rol del cuerpo se ignora. Sin cookie, `actor_rol` sigue valiendo
    para las labores (compatibilidad del piloto). No hay SSO.
    La agenda de reservas es un mock ficticio y no consulta hojas de cálculo.
servers:
  - url: http://127.0.0.1:8091
    description: Desarrollo local
tags:
  - name: sistema
  - name: catastro
  - name: operacion
paths:
  /health:
    get:
      tags: [sistema]
      summary: Salud del proceso, Postgres y PostGIS
      operationId: health
      responses:
        "200":
          description: Base y PostGIS disponibles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: Base o PostGIS no disponibles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /api/v1:
    get:
      tags: [sistema]
      summary: Índice de rutas
      operationId: indice
      responses:
        "200":
          description: Índice
          content:
            application/json:
              schema:
                type: object
                properties:
                  servicio:
                    type: string
                  version:
                    type: string
                  crs:
                    type: string
                    example: EPSG:4326
                  rutas:
                    type: array
                    items:
                      type: object
  /api/v1/openapi.yaml:
    get:
      tags: [sistema]
      summary: Este contrato
      operationId: openapi
      responses:
        "200":
          description: OpenAPI YAML
          content:
            application/yaml:
              schema:
                type: string
  /api/v1/geo/resumen:
    get:
      tags: [catastro]
      summary: Conteos de áreas, zonas y capas auxiliares
      operationId: resumenCatastro
      responses:
        "200":
          description: Conteos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resumen"
        "503":
          description: Base no disponible
  /api/v1/geo/areas:
    get:
      tags: [catastro]
      summary: Áreas verdes
      description: FeatureCollection de `areas_verdes`. Identificador estable `AV-NNNN`.
      operationId: areas
      parameters:
        - $ref: "#/components/parameters/bbox"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: FeatureCollection
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/FeatureCollection"
        "400":
          description: bbox o limit inválidos
        "503":
          description: Base no disponible
  /api/v1/geo/zonas:
    get:
      tags: [catastro]
      summary: Zonas de catastro
      description: |
        FeatureCollection derivada de `jefe_de_grupo.json`.
        El campo `jefes` (nombres de personas) no se almacena ni se sirve.
        Identificador estable `Z-NNNN`.
      operationId: zonas
      parameters:
        - $ref: "#/components/parameters/bbox"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: FeatureCollection
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/FeatureCollection"
        "400":
          description: bbox o limit inválidos
        "503":
          description: Base no disponible
  /api/v1/geo/capas:
    get:
      tags: [catastro]
      summary: Catálogo de capas auxiliares
      operationId: capas
      responses:
        "200":
          description: Capas conocidas y conteos cargados
          content:
            application/json:
              schema:
                type: object
                properties:
                  capas_conocidas:
                    type: array
                    items:
                      type: string
                  cargadas:
                    type: array
                    items:
                      $ref: "#/components/schemas/CapaCount"
  /api/v1/geo/capas/{capa}:
    get:
      tags: [catastro]
      summary: Capa auxiliar
      description: "`jardines_reserva` o `xerofitica`."
      operationId: capa
      parameters:
        - name: capa
          in: path
          required: true
          schema:
            type: string
            enum: [jardines_reserva, xerofitica]
        - $ref: "#/components/parameters/bbox"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: FeatureCollection
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/FeatureCollection"
        "404":
          description: Capa desconocida
        "503":
          description: Base no disponible
  /api/v1/operacion/actividades:
    get:
      tags: [operacion]
      summary: Labores como FeatureCollection (Point EPSG:4326)
      parameters:
        - in: query
          name: rol
          schema: { type: string, enum: [jefatura, coordinacion, capataz] }
        - in: query
          name: capataz_id
          schema: { type: string }
          description: Obligatorio si rol=capataz. Solo ve las asignadas a ese equipo.
        - in: query
          name: abiertas
          schema: { type: string, default: "1" }
          description: 1 excluye cerradas, canceladas y archivadas.
      responses:
        "200":
          description: FeatureCollection
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/FeatureCollection"
        "400":
          description: Filtro inválido
    post:
      tags: [operacion]
      summary: Crear labor desde un pin. El UUID del cliente es idempotente.
      responses:
        "201":
          description: Creada
        "200":
          description: El mismo UUID ya existía con el mismo contenido
        "403":
          description: El rol capataz no crea labores
        "409":
          description: UUID repetido con otro contenido
  /api/v1/geo/edificios:
    get:
      tags: [catastro]
      summary: Huellas OSM del recinto para extrusión
      responses:
        "200":
          description: FeatureCollection, vacía si falta el extracto
  /api/v1/geo/inventario:
    get:
      tags: [catastro]
      summary: Catálogo de overlays de inventario
      responses:
        "200":
          description: Capas conocidas y conteos cargados
  /api/v1/geo/inventario/{capa}:
    get:
      tags: [catastro]
      summary: Capa opcional de inventario
      parameters:
        - name: capa
          in: path
          required: true
          schema:
            type: string
            enum: [bebederos, fauna, playas_estacionamiento, puertas, area_vereda_peligro, flora, cafetos, tachos]
      responses:
        "200":
          description: FeatureCollection
        "404":
          description: Capa desconocida
  /api/v1/geo/reservas-mock:
    get:
      tags: [catastro]
      summary: Agenda ficticia, sin identificadores de hojas de cálculo
      responses:
        "200":
          description: Lista marcada como fake
  /api/v1/sesion:
    post:
      tags: [sistema]
      summary: Ingreso con cuenta local. Responde Set-Cookie HttpOnly
      responses:
        "200":
          description: Sesión abierta
        "401":
          description: Credenciales incorrectas
    get:
      tags: [sistema]
      summary: Usuario de la cookie
      responses:
        "200":
          description: Sesión vigente
        "401":
          description: Sin sesión
    delete:
      tags: [sistema]
      summary: Cierra la sesión
      responses:
        "200":
          description: Cookie retirada
  /api/v1/catalogos:
    get:
      tags: [sistema]
      summary: Ítems de catálogo. Query clase y activos=1
      responses:
        "200":
          description: Lista
    post:
      tags: [sistema]
      summary: Alta o reactivación. Requiere permiso catalogos
      responses:
        "201":
          description: Ítem guardado
  /api/v1/catastro/areas:
    get:
      tags: [catastro]
      summary: Fichas para editar metadatos
      responses:
        "200":
          description: Hasta 40 áreas
    post:
      tags: [catastro]
      summary: Crea un área sin geometría
      responses:
        "201":
          description: Ficha creada
  /api/v1/solicitudes:
    get:
      tags: [operacion]
      summary: Solicitudes con código externo conservado, no generado
      responses:
        "200":
          description: Lista
    post:
      tags: [operacion]
      summary: Alta manual (Centuria, OSG, correo o interna)
      responses:
        "201":
          description: Solicitud creada
  /api/v1/ordenes:
    post:
      tags: [operacion]
      summary: Orden de servicio para una labor tercerizada
      responses:
        "201":
          description: Orden creada
  /api/v1/riego:
    get:
      tags: [operacion]
      summary: Riego por sector y turno, sin indicador oficial
      responses:
        "200":
          description: Registros
  /api/v1/reportes/labores:
    get:
      tags: [operacion]
      summary: Reporte básico. formato=csv o formato=xls exporta el mismo filtro
      parameters:
        - name: formato
          in: query
          schema: { type: string, enum: [csv, xls] }
        - name: estado
          in: query
          schema: { type: string }
        - name: desde
          in: query
          schema: { type: string, example: "2026-09-01" }
        - name: hasta
          in: query
          schema: { type: string }
      responses:
        "200":
          description: JSON, CSV o SpreadsheetML
  /api/v1/ia/sugerir-tipo:
    post:
      tags: [operacion]
      summary: Regla local sobre el título. No llama a un modelo externo
      responses:
        "200":
          description: Sugerencia que una persona debe confirmar
components:
  parameters:
    bbox:
      name: bbox
      in: query
      required: false
      description: minLon,minLat,maxLon,maxLat en EPSG:4326. Filtra por intersección.
      schema:
        type: string
        example: "-77.09,-12.08,-77.07,-12.06"
    limit:
      name: limit
      in: query
      required: false
      description: Máximo de features (1–10000). Sin el parámetro se devuelve la colección completa.
      schema:
        type: integer
        minimum: 1
        maximum: 10000
  schemas:
    Health:
      type: object
      properties:
        status:
          type: string
          enum: [ok, error]
        service:
          type: string
        database:
          type: string
          enum: [up, down]
        postgis:
          type: string
    Resumen:
      type: object
      properties:
        crs:
          type: string
          example: EPSG:4326
        areas:
          type: integer
        areas_con_geometria:
          type: integer
        zonas:
          type: integer
        zonas_con_geometria:
          type: integer
        capas:
          type: array
          items:
            $ref: "#/components/schemas/CapaCount"
    CapaCount:
      type: object
      properties:
        capa:
          type: string
        features:
          type: integer
    FeatureCollection:
      type: object
      required: [type, features]
      properties:
        type:
          type: string
          enum: [FeatureCollection]
        name:
          type: string
        features:
          type: array
          items:
            $ref: "#/components/schemas/Feature"
    Feature:
      type: object
      required: [type, geometry, properties]
      properties:
        type:
          type: string
          enum: [Feature]
        id:
          type: string
          description: feature_id estable (AV-NNNN, Z-NNNN, JR-NNNN, XE-NNNN)
        geometry:
          nullable: true
          description: MultiPolygon EPSG:4326, o null si aún no hay GPS
        properties:
          type: object
          properties:
            id:
              type: integer
            feature_id:
              type: string
            source_index:
              type: integer
            codigo:
              type: string
            nombre:
              type: string
            uso:
              type: string
            proy_riego:
              type: string
            riego_act:
              type: string
            referencia:
              type: string
            perimetro_m:
              type: number
            area_m2:
              type: number
            capa:
              type: string
            clase:
              type: string
            pertenecen:
              type: string
              description: Unidad organizacional, no una persona